2_immune_cell_differentiation_rag/requirements.txt
# Immune Cell Differentiation RAPTOR Tree RAG System
# Requirements for Python packages

# Core ML/NLP Dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU available
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Graph and Network Analysis
networkx>=3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Document Processing
langchain>=0.1.0
langchain-core>=0.1.0

# API and Web
requests>=2.31.0
urllib3>=2.0.0

# Data Serialization
json5>=0.9.0

# Development and Utilities
tqdm>=4.65.0
python-dateutil>=2.8.0
pathlib2>=2.3.0  # For Python < 3.4 compatibility

# Optional: For advanced text processing
nltk>=3.8.0
spacy>=3.6.0

# Optional: For GPU acceleration
# faiss-gpu>=1.7.4  # Uncomment if using GPU
# torch-audio>=2.0.0  # For audio processing if needed

# Optional: For document parsing
# PyMuPDF>=1.23.0  # For PDF processing
# python-docx>=0.8.11  # For Word documents
# openpyxl>=3.1.0  # For Excel files

# Testing and Quality Assurance
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Optional: For advanced evaluation metrics
# rouge-score>=0.1.2
# bleu>=0.4.0
# sacrebleu>=2.3.0